+++ language-server/src/completer.rs
@@ -1,6 +1,6 @@
 use std::{collections::HashMap, sync::Arc};
 
-use ecow::EcoString;
+use ecow::{EcoString, eco_format};
 use itertools::Itertools;
 use lsp_types::{
     CompletionItem, CompletionItemKind, CompletionItemLabelDetails, CompletionTextEdit,
@@ -78,14 +78,182 @@ fn sort_text(kind: CompletionKind, label: &str, type_match: TypeMatch) -> String
     }
 }
 
-// The form in which a type completion is needed in context.
-// Mainly used to determine if the "type" keyword should be appended to the completion
-enum TypeCompletionForm {
-    // The type completion is for an unqualified import.
+/// The form in which a type completion is needed in context.
+#[derive(Debug)]
+enum TypeCompletionContext {
+    /// The type completion is for an unqualified import that doesn't have an
+    /// import list yet. So adding the type will also require adding braces:
+    ///
+    /// ```gleam
+    /// import wibble.Wib|
+    /// //           ^^^^^ We're typing this...
+    /// import wibble.{type Wibble}
+    /// //           ^^^^^^^^^^^^^^ ...so this will be the completion
+    /// ```
+    ///
     UnqualifiedImport,
-    // The type completion is for an unqualified import within braces.
+
+    /// The type completion is for an unqualified import within already existing
+    /// braces.
+    ///
+    /// ```gleam
+    /// import wibble.{type AlreadyImported, Wibb|}
+    /// //                                   ^^^^^ We're typing this...
+    /// import wibble.{type AlreadyImported, type Wibble}
+    /// //                                   ^^^^^^^^^^^ ...so this will be the completion
+    /// ```
+    ///
     UnqualifiedImportWithinBraces,
-    Default,
+
+    /// The type completion is for an unqualified type (for example something
+    /// coming from the prelude, or a type that was already imported in a
+    /// unqualified way).
+    ///
+    /// ```gleam
+    /// import wobble.{type Wobble}
+    ///
+    /// pub fn new_wobble() -> Wob|
+    /// //                     ^^^^ We're typing this...
+    /// pub fn new_wobble() -> Wobble
+    /// //                     ^^^^^^ ... so this will be the completion
+    /// ```
+    ///
+    UnqualifiedType,
+
+    /// The type completion is for a qualified type.
+    ///
+    /// ```gleam
+    /// import wobble
+    ///
+    /// pub fn new_wobble() -> wobble.W|
+    /// //                     ^^^^^^^^ We're typing this...
+    /// pub fn new_wobble() -> wobble.Wobble
+    /// //                     ^^^^^^^^^^^^^ ...so this will be the completion
+    /// ```
+    ///
+    QualifiedType,
+}
+
+/// Represents the surroundings of the cursor when trying to figure out a
+/// completion.
+///
+/// ```gleam
+/// wibble.w|ob
+/// //      ^ cursor here
+/// ```
+struct CursorSurroundings {
+    /// The text surrounding the cursor. For example:
+    ///
+    /// ```gleam
+    /// wibble.w|ob
+    /// //      ^ cursor here
+    /// // The surrounding text is: "wibble.wob"
+    /// ```
+    ///
+    surrounding_text: EcoString,
+    surrounding_text_range: Range,
+
+    /// The text that comes immediately before the cursor.
+    /// For example:
+    ///
+    /// ```gleam
+    ///    wibble.w|ob
+    /// //         ^ cursor here
+    /// // The text before is: "wibble.w"
+    /// ```
+    ///
+    text_before_cursor: EcoString,
+
+    /// The range of the text that comes immediately before the cursor.
+    /// For example:
+    ///
+    /// ```gleam
+    ///    wibble.w|ob
+    /// //         ^ cursor here
+    /// // ^^^^^^^^ This is the range
+    /// ```
+    ///
+    /// This is what is usually replaced with a completion.
+    ///
+    text_before_cursor_range: Range,
+
+    /// The text that comes immediately after the cursor.
+    /// For example:
+    ///
+    /// ```gleam
+    ///    wibble.w|ob
+    /// //         ^ cursor here
+    /// // The text before is: "ob"
+    /// ```
+    ///
+    text_after_cursor: EcoString,
+}
+
+impl CursorSurroundings {
+    fn selected_module(&self) -> Option<EcoString> {
+        self.surrounding_text
+            .split_once('.')
+            .map(|(selected_module, _)| EcoString::from(selected_module))
+    }
+
+    /// Given a proposed completion, this returns the text edit to obtain that
+    /// completion.
+    ///
+    /// > This could also return `None` as sometimes no further edit is actually
+    /// > needed!
+    ///
+    fn to_text_edit(&self, new_text: String) -> Option<CompletionTextEdit> {
+        // We need to check if the new text we're adding could actually be
+        // a simple addition in the middle of something that is already being
+        // typed.
+        // Say we're editing our code to actually make the `Json` type
+        // qualified:
+        //
+        // ```gleam
+        // jso|Json
+        //    ^ cursor here
+        // ```
+        //
+        // Halfway through writing the module name we might just want to accept
+        // the completion for: `json.Json`.
+        // What one would normally expect is for the final code to be:
+        //
+        // ```gleam
+        // json.Json|
+        // // after accepting completion
+        //
+        // // and not something like this!
+        // // json.JsonJson
+        // ```
+        //
+        // This only makes sense if the completion we're adding has a prefix and
+        // suffix in common with the surrounding text:
+        let remaining_label = new_text
+            .strip_prefix(self.text_before_cursor.as_str())
+            .and_then(|rest| rest.strip_suffix(self.text_after_cursor.as_str()));
+
+        match remaining_label {
+            // The entire existing text is already the same as the new text we
+            // might want to add.
+            // In that case there's no need to perform any edit at all!
+            Some("") => None,
+
+            // This is one of the cases (like the one in the example above) were
+            // the label is a more complete version of the text surrounding the
+            // cursor. So we replace the entire range.
+            Some(_) => Some(CompletionTextEdit::Edit(TextEdit {
+                range: self.surrounding_text_range,
+                new_text,
+            })),
+            // In all other cases the completion is never meant to replace text
+            // that comes after the cursor.
+            // We only replace what comes before it with the new text.
+            None => Some(CompletionTextEdit::Edit(TextEdit {
+                range: self.text_before_cursor_range,
+                new_text,
+            })),
+        }
+    }
 }
 
 pub struct Completer<'a, IO> {
@@ -135,7 +303,7 @@ impl<'a, IO> Completer<'a, IO> {
     fn get_phrase_surrounding_for_completion(
         &'a self,
         valid_phrase_char: &impl Fn(char) -> bool,
-    ) -> (Range, String) {
+    ) -> CursorSurroundings {
         let cursor = self.src_line_numbers.byte_index(*self.cursor_position);
 
         // Get part of phrase prior to cursor
@@ -152,45 +320,54 @@ impl<'a, IO> Completer<'a, IO> {
             .and_then(|line| line.split_once(valid_phrase_char).map(|r| r.0))
             .unwrap_or("");
 
-        (
-            Range {
-                start: Position {
-                    line: self.cursor_position.line,
-                    character: self.cursor_position.character - before.len() as u32,
-                },
-                end: Position {
-                    line: self.cursor_position.line,
-                    character: self.cursor_position.character,
-                },
+        let text_before_cursor_range = Range {
+            start: Position {
+                line: self.cursor_position.line,
+                character: self.cursor_position.character - before.len() as u32,
             },
-            format!("{before}{after}"),
-        )
+            end: Position {
+                line: self.cursor_position.line,
+                character: self.cursor_position.character,
+            },
+        };
+
+        let surrounding_text_range = Range {
+            start: text_before_cursor_range.start,
+            end: Position {
+                line: self.cursor_position.line,
+                character: self.cursor_position.character + after.len() as u32,
+            },
+        };
+
+        CursorSurroundings {
+            surrounding_text: eco_format!("{before}{after}"),
+            surrounding_text_range,
+            text_before_cursor: EcoString::from(before),
+            text_before_cursor_range,
+            text_after_cursor: EcoString::from(after),
+        }
     }
 
     // Gets the current range around the cursor to place a completion
     // and any part of the phrase preceeding a dot if a module is being selected from.
     // A continuous phrase in this case is a name or typename that may have a dot in it.
     // This is used to match the exact location to fill in the completion.
-    fn get_phrase_surrounding_completion(&'a self) -> (Range, Option<String>) {
-        let valid_phrase_char = |c: char| {
+    fn get_phrase_surrounding_completion(&'a self) -> CursorSurroundings {
+        self.get_phrase_surrounding_for_completion(&|c: char| {
             // Checks if a character is not a valid name/upname character or a dot.
             !c.is_ascii_alphanumeric() && c != '.' && c != '_'
-        };
-        let (range, word) = self.get_phrase_surrounding_for_completion(&valid_phrase_char);
-        (range, word.split_once('.').map(|c| String::from(c.0)))
+        })
     }
 
     // Gets the current range around the cursor to place a completion.
     // For unqualified imports we special case the word being completed to allow for whitespace but not dots.
     // This is to allow `type MyType` to be treated as 1 "phrase" for the sake of completion.
-    fn get_phrase_surrounding_completion_for_import(&'a self) -> Range {
-        let valid_phrase_char = |c: char| {
+    fn get_phrase_surrounding_completion_for_import(&'a self) -> CursorSurroundings {
+        self.get_phrase_surrounding_for_completion(&|c: char| {
             // Checks if a character is not a valid name/upname character or whitespace.
             // The newline character is not included as well.
             !c.is_ascii_alphanumeric() && c != '_' && c != ' ' && c != '\t'
-        };
-        let (range, _) = self.get_phrase_surrounding_for_completion(&valid_phrase_char);
-        range
+        })
     }
 
     /// Checks if the line being editted is an import line and provides completions if it is.
@@ -249,7 +426,7 @@ impl<'a, IO> Completer<'a, IO> {
         module_being_imported_from: &'a ModuleInterface,
         within_braces: bool,
     ) -> Vec<CompletionItem> {
-        let insert_range = self.get_phrase_surrounding_completion_for_import();
+        let cursor_surroundings = self.get_phrase_surrounding_completion_for_import();
         let mut completions = vec![];
 
         // Find values and type that have already previously been imported
@@ -290,11 +467,11 @@ impl<'a, IO> Completer<'a, IO> {
                 None,
                 name,
                 type_,
-                insert_range,
+                &cursor_surroundings,
                 if within_braces {
-                    TypeCompletionForm::UnqualifiedImportWithinBraces
+                    TypeCompletionContext::UnqualifiedImportWithinBraces
                 } else {
-                    TypeCompletionForm::UnqualifiedImport
+                    TypeCompletionContext::UnqualifiedImport
                 },
                 CompletionKind::ImportedModule,
             ));
@@ -319,7 +496,7 @@ impl<'a, IO> Completer<'a, IO> {
                 &module_being_imported_from.name,
                 name,
                 value,
-                insert_range,
+                &cursor_surroundings,
                 CompletionKind::ImportedModule,
             ));
         }
@@ -410,23 +587,23 @@ impl<'a, IO> Completer<'a, IO> {
 
     /// Provides completions for when the context being editted is a type.
     pub fn completion_types(&'a self) -> Vec<CompletionItem> {
-        let surrounding_completion = self.get_phrase_surrounding_completion();
+        let cursor_surroundings = self.get_phrase_surrounding_completion();
+        let selected_module = cursor_surroundings.selected_module();
         let mut completions = vec![];
 
-        let (insert_range, module_select) = surrounding_completion;
-
         // Module and prelude types
-        // Do not complete direct module types if the user has already started typing a module select.
-        // e.x. when the user has typed mymodule.| we know local module types and prelude types are no
-        // longer relevant.
-        if module_select.is_none() {
+        // Do not complete direct module types if the user has already started
+        // typing a module select. That is, when the user has already typed
+        // `mymodule.|` we know local module types and prelude types are no
+        // longer relevant and needed for completions.
+        if selected_module.is_none() {
             for (name, type_) in &self.module.ast.type_info.types {
                 completions.push(type_completion(
                     None,
                     name,
                     type_,
-                    insert_range,
-                    TypeCompletionForm::Default,
+                    &cursor_surroundings,
+                    TypeCompletionContext::UnqualifiedType,
                     CompletionKind::LocallyDefined,
                 ));
             }
@@ -448,51 +625,56 @@ impl<'a, IO> Completer<'a, IO> {
             }
         }
 
-        // Imported modules
+        // Qualified types
         for import in &self.module.ast.definitions.imports {
             // The module may not be known of yet if it has not previously
             // compiled yet in this editor session.
             let Some(module) = self.compiler.get_module_interface(&import.module) else {
                 continue;
             };
+            let Some(module_name) = &import.used_name() else {
+                continue;
+            };
 
-            // Qualified types
             for (name, type_) in &module.types {
                 if !self.is_suggestable_import(&type_.publicity, module.package.as_str()) {
                     continue;
                 }
 
-                if let Some(module) = import.used_name() {
-                    // If the user has already started a module select then don't show irrelevant modules.
-                    // e.x. when the user has typed mymodule.| we should only show items from mymodule.
-                    if let Some(input_mod_name) = &module_select
-                        && &module != input_mod_name
-                    {
-                        continue;
-                    }
-                    completions.push(type_completion(
-                        Some(&module),
-                        name,
-                        type_,
-                        insert_range,
-                        TypeCompletionForm::Default,
-                        CompletionKind::ImportedModule,
-                    ));
+                // If the user has already started typing a module select
+                // then don't show irrelevant modules.
+                // For example: when the user has typed `mymodule.|` we
+                // should only show items from `mymodule`.
+                if let Some(typed_module) = &selected_module
+                    && module_name != typed_module
+                {
+                    continue;
                 }
+
+                completions.push(type_completion(
+                    Some(module_name),
+                    name,
+                    type_,
+                    &cursor_surroundings,
+                    TypeCompletionContext::QualifiedType,
+                    CompletionKind::ImportedModule,
+                ));
             }
 
             // Unqualified types
-            // Do not complete unqualified types if the user has already started typing a module select.
-            // e.x. when the user has typed mymodule.| we know unqualified module types are no longer relevant.
-            if module_select.is_none() {
+            // Do not complete unqualified types if the user has already started
+            // typing a module select.
+            // For example, when the user has already typed `mymodule.|` we know
+            // unqualified module types are no longer relevant.
+            if selected_module.is_none() {
                 for unqualified in &import.unqualified_types {
                     if let Some(type_) = module.get_public_type(&unqualified.name) {
                         completions.push(type_completion(
                             None,
                             unqualified.used_name(),
                             type_,
-                            insert_range,
-                            TypeCompletionForm::Default,
+                            &cursor_surroundings,
+                            TypeCompletionContext::UnqualifiedType,
                             CompletionKind::ImportedModule,
                         ))
                     }
@@ -525,8 +707,8 @@ impl<'a, IO> Completer<'a, IO> {
 
             // If the user has already started a module select then don't show irrelevant modules.
             // e.x. when the user has typed mymodule.| we should only show items from mymodule.
-            if let Some(input_mod_name) = &module_select
-                && qualifier != input_mod_name
+            if let Some(selected_module) = &selected_module
+                && qualifier != selected_module
             {
                 continue;
             }
@@ -541,8 +723,8 @@ impl<'a, IO> Completer<'a, IO> {
                     Some(qualifier),
                     name,
                     type_,
-                    insert_range,
-                    TypeCompletionForm::Default,
+                    &cursor_surroundings,
+                    TypeCompletionContext::QualifiedType,
                     CompletionKind::ImportableModule,
                 );
                 add_import_to_completion(
@@ -560,17 +742,16 @@ impl<'a, IO> Completer<'a, IO> {
 
     /// Provides completions for when the context being editted is a value.
     pub fn completion_values(&'a self) -> Vec<CompletionItem> {
-        let surrounding_completion = self.get_phrase_surrounding_completion();
+        let cursor_surroundings = self.get_phrase_surrounding_completion();
+        let selected_module = cursor_surroundings.selected_module();
         let mut completions = vec![];
         let mod_name = self.module.name.as_str();
-
-        let (insert_range, module_select) = surrounding_completion;
 
         // Module and prelude values
         // Do not complete direct module values if the user has already started typing a module select.
         // e.x. when the user has typed mymodule.| we know local module and prelude values are no longer
         // relevant.
-        if module_select.is_none() {
+        if selected_module.is_none() {
             let cursor = self.src_line_numbers.byte_index(*self.cursor_position);
 
             // Find the function that the cursor is in and push completions for
@@ -588,7 +769,7 @@ impl<'a, IO> Completer<'a, IO> {
                 completions.extend(
                     LocalCompletion::new(
                         mod_name,
-                        insert_range,
+                        cursor_surroundings.surrounding_text_range,
                         cursor,
                         self.expected_type.clone(),
                     )
@@ -605,7 +786,7 @@ impl<'a, IO> Completer<'a, IO> {
                     mod_name,
                     name,
                     value,
-                    insert_range,
+                    &cursor_surroundings,
                     CompletionKind::LocallyDefined,
                 ));
             }
@@ -686,7 +867,7 @@ impl<'a, IO> Completer<'a, IO> {
                 if let Some(module) = import.used_name() {
                     // If the user has already started a module select then don't show irrelevant modules.
                     // e.x. when the user has typed mymodule.| we should only show items from mymodule.
-                    if let Some(input_mod_name) = &module_select
+                    if let Some(input_mod_name) = &selected_module
                         && &module != input_mod_name
                     {
                         continue;
@@ -696,7 +877,7 @@ impl<'a, IO> Completer<'a, IO> {
                         mod_name,
                         name,
                         value,
-                        insert_range,
+                        &cursor_surroundings,
                         CompletionKind::ImportedModule,
                     ));
                 }
@@ -705,7 +886,7 @@ impl<'a, IO> Completer<'a, IO> {
             // Unqualified values
             // Do not complete unqualified values if the user has already started typing a module select.
             // e.x. when the user has typed mymodule.| we know unqualified module values are no longer relevant.
-            if module_select.is_none() {
+            if selected_module.is_none() {
                 for unqualified in &import.unqualified_values {
                     if let Some(value) = module.get_public_value(&unqualified.name) {
                         let name = unqualified.used_name();
@@ -714,7 +895,7 @@ impl<'a, IO> Completer<'a, IO> {
                             mod_name,
                             name,
                             value,
-                            insert_range,
+                            &cursor_surroundings,
                             CompletionKind::ImportedModule,
                         ))
                     }
@@ -745,8 +926,8 @@ impl<'a, IO> Completer<'a, IO> {
 
             // If the user has already started a module select then don't show irrelevant modules.
             // e.x. when the user has typed mymodule.| we should only show items from mymodule.
-            if let Some(input_mod_name) = &module_select
-                && qualifier != input_mod_name
+            if let Some(selected_module) = &selected_module
+                && qualifier != selected_module
             {
                 continue;
             }
@@ -762,7 +943,7 @@ impl<'a, IO> Completer<'a, IO> {
                     module_full_name,
                     name,
                     value,
-                    insert_range,
+                    &cursor_surroundings,
                     CompletionKind::ImportableModule,
                 );
 
@@ -922,7 +1103,7 @@ impl<'a, IO> Completer<'a, IO> {
         module_name: &str,
         name: &str,
         value: &type_::ValueConstructor,
-        insert_range: Range,
+        cursor_surrounding: &CursorSurroundings,
         priority: CompletionKind,
     ) -> CompletionItem {
         let type_match = match_type(&self.expected_type, &value.type_);
@@ -941,10 +1122,10 @@ impl<'a, IO> Completer<'a, IO> {
             ValueConstructorVariant::Record { .. } => CompletionItemKind::CONSTRUCTOR,
         });
 
-        let documentation = value.get_documentation().map(|d| {
+        let documentation = value.get_documentation().map(|documentation| {
             Documentation::MarkupContent(MarkupContent {
                 kind: MarkupKind::Markdown,
-                value: d.into(),
+                value: documentation.into(),
             })
         });
 
@@ -958,10 +1139,7 @@ impl<'a, IO> Completer<'a, IO> {
             }),
             documentation,
             sort_text: Some(sort_text(priority, &label, type_match)),
-            text_edit: Some(CompletionTextEdit::Edit(TextEdit {
-                range: insert_range,
-                new_text: label.clone(),
-            })),
+            text_edit: cursor_surrounding.to_text_edit(label),
             ..Default::default()
         }
     }
@@ -997,8 +1175,8 @@ fn type_completion(
     module: Option<&str>,
     name: &str,
     type_: &TypeConstructor,
-    insert_range: Range,
-    include_type_in_completion: TypeCompletionForm,
+    cursor_surrounding: &CursorSurroundings,
+    type_completion_context: TypeCompletionContext,
     priority: CompletionKind,
 ) -> CompletionItem {
     let label = match module {
@@ -1012,19 +1190,20 @@ fn type_completion(
         CompletionItemKind::CLASS
     });
 
+    let completion_text = match type_completion_context {
+        TypeCompletionContext::UnqualifiedImport => format!("{{type {label}}}"),
+        TypeCompletionContext::UnqualifiedImportWithinBraces => format!("type {label}"),
+        TypeCompletionContext::QualifiedType | TypeCompletionContext::UnqualifiedType => {
+            label.clone()
+        }
+    };
+
     CompletionItem {
         label: label.clone(),
         kind,
         detail: Some("Type".into()),
         sort_text: Some(sort_text(priority, &label, TypeMatch::Unknown)),
-        text_edit: Some(CompletionTextEdit::Edit(TextEdit {
-            range: insert_range,
-            new_text: match include_type_in_completion {
-                TypeCompletionForm::UnqualifiedImport => format!("{{type {label}}}"),
-                TypeCompletionForm::UnqualifiedImportWithinBraces => format!("type {label}"),
-                TypeCompletionForm::Default => label.clone(),
-            },
-        })),
+        text_edit: cursor_surrounding.to_text_edit(completion_text),
         ..Default::default()
     }
 }