+++ language-server/src/tests/completion.rs
@@ -161,8 +161,8 @@ fn format_completion_results(completions: Vec<CompletionItem>) -> EcoString {
         let edit = |buffer: &mut EcoString, e: lsp_types::TextEdit| {
             let a = e.range.start.line;
             let b = e.range.start.character;
-            let c = e.range.start.line;
-            let d = e.range.start.character;
+            let c = e.range.end.line;
+            let d = e.range.end.character;
             write!(buffer, "\n    [{a}:{b}-{c}:{d}]: {:?}", e.new_text).unwrap();
         };
 
@@ -2293,4 +2293,31 @@ pub fn main() -> Result(Int, Nil) {
 ";
 
     assert_completion!(TestProject::for_source(code), Position::new(4, 3));
+}
+
+#[test]
+fn completion_for_type() {
+    let dep = "pub type Wibble";
+    let code = "pub fn new() -> wibble.Wibble {}";
+    assert_completion!(
+        TestProject::for_source(code).add_dep_module("wibble", dep),
+        Position::new(0, 23)
+    );
+}
+
+#[test]
+fn completion_for_partially_correct_existing_module_select() {
+    let dep = "pub fn filter() {}";
+    let code = "
+import gleam/list
+
+pub fn new() {
+  list.fer
+//      ^ cursor right after the 'f'
+}
+";
+    assert_completion!(
+        TestProject::for_source(code).add_dep_module("gleam/list", dep),
+        Position::new(4, 8)
+    );
 }